Desktop/password-manager/requirements.txt
alabaster==1.0.0
altair==5.2.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==23.2.0
babel==2.16.0
beautifulsoup4==4.13.2
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decouple==0.0.7
Django==5.1.7
docutils==0.21.2
dotenv==0.9.9
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.17.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.11
GitPython==3.1.42
glib==1.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.6
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.8
lap==0.5.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
meson==1.6.1
mpmath==1.3.0
mysql-connector-python==9.2.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==23.2
pandas==2.2.1
pillow==10.2.0
plotly==5.20.0
protobuf==4.25.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==15.0.2
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.34.0
regex==2024.11.6
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==75.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
starlette==0.46.2
streamlit==1.32.2
sympy==1.13.1
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4
tqdm==4.67.1
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
ultralytics==8.3.85
ultralytics-thop==2.0.14
urllib3==2.2.1
uvicorn==0.34.1
watchdog==4.0.0
Werkzeug==3.1.3
